Rename directory with React components from Componets to Components
